Moved functions that create HTML elements into a single file: elementCreation.js. #5
